requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
cloudinary==1.40.0
dj-database-url==2.1.0
Django==5.0.6
django-ckeditor==6.7.1
django-cloudinary-storage==0.3.0
django-environ==0.11.2
django-heroku==0.3.1
django-js-asset==2.2.0
gunicorn==22.0.0
idna==3.7
packaging==24.0
pillow==10.3.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
requests==2.32.2
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
whitenoise==6.6.0
